Fixed the bug of "cleared dict data"!! :-) The problem was that we opened the root file to get the Nevents from the histo, than we left the file open and we chained it with t.Chain(file). Then the file was automaticcaly closed somewhere else at the end of some loops and the TChain lost its handle to the file!!
so the solution was: open the file, read the Nevents histo, then close the file and then chain it with TChain.

In this way TChain keeps an handle to the file even if it's closed.

P.s. I also updated some debug messages and on/off switches for sections of the code.