Store muon trigger features independent of muons, and tried to compactify MS track summary information.
The L1 ROI, and the L2 and EF muon features are now stored independent of the muons. The muons
only contain an index and the delta R of the matching. 
The per layer hit information for TGCs and RPCs is now stored in vectors instead of using one
branch per layer.

For the association of muons and trigger features, the TriggerContainerGetterTool had to be corrected
Keep a private corrected version until the version from TriggerD3PDMaker handles index association.