Added cuts on muon energy and energy loss to reduce number of fake calo muons.
ADD: Cuts on energy and energy loss, which can be used to reduce the number of fake calo muons.

Moreover tried to remove text comparisons from execute i.e. muon author and type which are passed
as string to the tool are parsed in initialize and turned into an enum.